requirements.txt
requests==2.32.3
geopandas==0.13.2
folium==0.17.0
shapely==2.0.5
pytz==2024.1
tqdm==4.66.5
python-dotenv=1.0.1